+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.7.3 () - 'Scanners'
 
-- New: option to reply to the original post instead of the reposted one
+- New: option to reply to the original post if replying to a reposted post. Example: `ayadn -R --noredirect 123456`. (default: reply to reposted)
+- New: you can change the users lists order. Example: `ayadn set formats list reverse false` (default: true)
+- New: you can change the table width. Example: `ayadn set formats table width 80` (min: 60, max: 90, default: 75)
 
 # 1.7.2 (2014-09-04) - 'Fortress'